Misc fixes all over the place; shellcheck fixes mostly.
checksum_file.sh
* explained the OpenBSD situation better
* improved print_usage()
* moved OS variants into alphabetical order
* several shellcheck fixes

rsnapshot/rsnapshot-rotate.sh
* several shellcheck fixes
* set -e
* omit WRAPPER_CONF

+++ ternet/flop.sh
@@ -5,15 +5,17 @@
 # https://web.archive.org/web/20170702162404/http://www.breakingsystemsforfunandprofit.com/the-lolcats-they-are-everywhere/
 #
 DIR=/var/www/ternet
-while read URL; do
+
+while read -r URL; do
 	SURL=$(echo "${URL}" | cut -d" " -f1)
-	echo "${SURL}" | egrep -qi "\.(jp(e)?g|gif|png|tiff|bmp|ico)$" &&
-	(
-	umask 002
-	PIC=$$-${RANDOM}
-	wget -q -O ${DIR}/${PIC}.tmp "${SURL}"
-	convert -quiet ${DIR}/${PIC}.tmp -flop ${DIR}/${PIC}.png
-	rm -f ${DIR}/${PIC}.tmp
-	echo http://127.0.0.1/ternet/${PIC}.png
-	) || echo "$URL"
+	if echo "${SURL}" | grep -Eqi "\.(jp(e)?g|gif|png|tiff|bmp|ico)$"; then
+		umask 002
+		PIC=$$-${RANDOM}
+		wget -q -O ${DIR}/${PIC}.tmp "${SURL}"
+		convert -quiet ${DIR}/${PIC}.tmp -flop ${DIR}/${PIC}.png
+		rm -f ${DIR}/${PIC}.tmp
+		echo http://127.0.0.1/ternet/${PIC}.png
+	else
+		echo "${URL}"
+	fi
 done
